Move controller related utils
Because they are used by tests for controllers, move them under the
controller directory.
