+++ internal/controller/internal/testutil/manager_util.go
@@ -1,4 +1,4 @@
-package util
+package testutil
 
 import (
 	"context"
