+++ test/util/util.go
@@ -1,13 +1,8 @@
 package util
 
 import (
-	"context"
 	"fmt"
 	"math/rand"
-
-	storagev1 "k8s.io/api/storage/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 var (
@@ -26,17 +21,4 @@ func GetUniqueName(prefix string) string {
 	}
 	usedResourceNames[name] = true
 	return name
-}
-
-func CreateStorageClass(ctx context.Context, client client.Client, name, provisioner, cephClusterID string) error {
-	sc := storagev1.StorageClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Provisioner: provisioner,
-		Parameters: map[string]string{
-			"clusterID": cephClusterID,
-		},
-	}
-	return client.Create(ctx, &sc)
 }
